requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
pydantic>=2.7.0
pydantic-settings>=2.2.0
bcrypt>=4.0.0
PyJWT>=2.8.0